refactor: extract browser + tab routes, deduplicate /tabs/source
- Move 14 browser routes to routes/browser.ts
- Move 7 tab routes to routes/tabs.ts
- Add buildContext() method and route registration
- Remove duplicate /tabs/source (was at two locations)
